ref: Split replay integration & container class
