+++ packages/replay/test/unit/index-handleGlobalEvent.test.ts
@@ -1,14 +1,14 @@
 import { getCurrentHub } from '@sentry/core';
 
 import { REPLAY_EVENT_NAME } from '../../src/constants';
-import { Replay } from './../../src';
+import { ReplayContainer } from './../../src/replay';
 import { Error } from './../fixtures/error';
 import { Transaction } from './../fixtures/transaction';
 import { resetSdkMock } from './../mocks';
 import { useFakeTimers } from './../utils/use-fake-timers';
 
 useFakeTimers();
-let replay: Replay;
+let replay: ReplayContainer;
 
 beforeEach(async () => {
   ({ replay } = await resetSdkMock({
